delete old node_monitor and update everything to the new version, making large breaking changes
