+++ Dockerfile
@@ -8,4 +8,12 @@ COPY . /app
 
 WORKDIR /app
 
-CMD ["python3", "-m", "node_monitor"]
+EXPOSE 80
+
+RUN python3 node_monitor/load_config.py
+
+# TODO: Should we put the check and test in a separate container?
+RUN make check
+RUN make test
+
+CMD ["make", "prod"]
