+++ README.md
@@ -1,59 +1,15 @@
-# Node Monitor by Aviate Labs
+# Node Monitor
 
 Node monitoring software for notifications of changes to Internet Computer Nodes.
-
-Queries the API with a specified interval (configurable) and reports changes to email.
-Can be configured to filter different kinds of events.
-
-Includes an optional feature to send a status update to any incoming emails that contain the word 'status' in the body or subject.
-
+Queries the API with a specified interval and reports changes to email and other various communication channels.
 
 ## Setup
-Create a .env file in this directory like so
-```text
-gmailUsername       = "email.sender@gmail.com"
-gmailPassword       = "mypassword"
-slackBotToken       = "your-slack-bot-api-token"
-telegramBotToken    = "your-telegram-bot-api-token"
-```
-The email address listed in the `gmailUsername` field serves as the primary sender for notifications and the point of contact for getting the status of your nodes. 
 
-If you have 2FA enabled on your gmail account. You will need to use an application specific password in the `gmailPassword` field. Follow the steps [here](https://support.google.com/mail/answer/185833?hl=en-GB).
+Place a `.env` file in this directory.  
+Use `.env.example` as a template.
 
-
-Create a config.json file in this directory. Below is a good default config.
-```js
-{
-    "emailRecipients": ["email.receiver1@gmail.com", "email.receiver2@gmail.com"],
-    "slackChannelName": "your-node-monitor-channel",
-    "telegramChatId": "0123456789",
-    "telegramChannelId": "-1234567890987",
-    "nodeProviderId": "abc2d-48fgj-32ab3-2a...",
-    "intervalMinutes": 5,
-    "intervalStatusReport": 1440, 
-    "lookupTableFile": "lookuptable.json",
-    "NotifyOnNodeMonitorStartup": true,
-    "NotifyOnNodeChangeStatus": true,
-    "NotifyOnAllNodeChanges": false,
-    "NotifyOnNodeAdded": true,
-    "NotifyOnNodeRemoved": true,
-    "IMAPClientEnabled": false,
-    "NotifyBySlack": true,
-    "NotifyByEmail": true
-}
-```
-The `intervalStatusReport` value, set in minutes, determines how often you'll receive these reports. It should always be larger than the `intervalMinutes` value. If it this is not set in your config.json file, or the `intervalStatusReport` < `intervalMinutes`, you will not recieve a status report.
-
-The lookuptable.json file is mapping of node-ids to custom labels. Its use is optional.
-Example lookuptable.json:
-```js
-{
-    "5jo56-bx...": "an1-dll01",
-    "vseni-vz...": "an1-dll02",
-    "rbr7v-bi...": "an1-dll03",
-    "kby7l-tp...": "an1-dll04",
-}
-```
+You will also need a file `tmp_prod_db.json` in this directory.  
+This is our temporary database until we implement SQLite.
 
 
 Install dependencies
@@ -62,59 +18,47 @@ $ pip install -r requirements.txt
 ```
 
 
-## Running
-Run with python, and stop with keyboardinterrupt (crtl-c)
-```sh
-$ python3 -m node_monitor
-[2022-10-08 07:52:37.210304]: Starting Node Monitor...
-[2022-10-08 07:52:37.982442]: -- Fetched New Data
-[2022-10-08 07:52:38.650524]: -- Fetched New Data
-[2022-10-08 07:52:38.671211]: -- No Change (5 min)
-^C[2022-10-08 07:52:41.509254]: Stopped Node Monitor
+## Testing
+
+We use mypy and pytest to do testing and type checking.
+
+```bash
+$ make check
+$ make test
+# To run the test with a live email
+# -s flag is to show print statements
+$ pytest -s --send_emails tests/
 ```
 
 
-## TODO
-- Better error handling
 
-## Using with rsync
+## Running
+
+We use Gunicorn as our WSGI server.
+
+```bash
+$ make prod
+```
+
+
+## Notes
+
+#### Usage with rsync
 ```bash
 # a=preserve links, n=dry run, v=verbose
-$ rsync -anv --exclude 'config.json' --exclude '.git/' . username@remote_host:/root/directory
+$ rsync -anv --exclude '.git/' . username@remote_host:/root/directory
 # remove the `n` to run for real
 ```
 
-## Testing
-```sh
-$ python3 -m unittest tests/test_node_monitor.py
-```
-### Notes
+#### 2FA Gmail
+If you have 2FA enabled on your gmail account. You will need to use an application specific password in the `gmailPassword` field. Follow the steps [here](https://support.google.com/mail/answer/185833?hl=en-GB).
+
+#### Fetching data from the API
 For testing, fetch api data with curl. The repo should already include tests and accompanying json documents, so this shouldn't really be necessary
 ```sh
 $ curl "https://ic-api.internetcomputer.org/api/v3/nodes" -o t0.json
 ```
 
-
-
-## Settings Explained
-
-- intervalMinutes: The interval in which to wait to query the dashboard API
-
-- NotifyOnNodeMonitorStartup: Send a notification email each time Node Monitor is started
-
-- NotifyOnNodeChangeStatus: Sends an email if node changes to 'UP', 'DOWN', or 'UNASSIGNED'
-
-- NotifyOnAllNodeChanges: Sends an email for any kind of node change (subnet, owner, etc), basically verbose mode
-
-- NotifyOnNodeAdded: Send an email when a new node is discovered on the dashboard API
-
-- NotifyOnNodeRemoved: Send an email when a node is removed from the dashboard API
-
-- IMAPClientEnabled: Watch inbox and send any responses to emails querying a status update
-
-- NotfyBySlack: Sends Slack messages according to the settings specified above
-
-- NotifyByTelegram: Sends Telegram messages according to the settings specified above
 
 ## Set up Node Monitor Slack Bot
 
@@ -140,3 +84,4 @@ $ curl "https://ic-api.internetcomputer.org/api/v3/nodes" -o t0.json
     ```
 
 - Now you have the bot API token and the chat ID, put the bot API token in the `.env` file in the `telegramBotToken` field and the chat ID in the `config.json` file in the `telegramChatId` field.
+
