--- minified/node_monitor/load_config.py
@@ -1,13 +1,0 @@
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-
-##############################################
-## Secrets
-
-gmailUsername    = os.environ.get('gmailUsername',   '')
-gmailPassword    = os.environ.get('gmailPassword',   '')
-discordBotToken  = os.environ.get('discordBotToken', '')  # Not implemented
-slackBotToken    = os.environ.get('slackBotToken',   '')
