+++ node_monitor/node_monitor_helpers/get_compromised_nodes.py
