+++ requirements.txt
@@ -1,6 +1,11 @@
-deepdiff
-deepdiff6
-discord.py
-python-dotenv
+Flask
+gunicorn
+toolz
 requests
-slackclient
+pydantic
+python-dotenv
+mypy
+types-requests
+slack_sdk
+devtools
+pytest
