+++ tests/node_monitor_helpers/test_get_compromised_nodes.py
